Add Hana support for .NetCore
