+++ dotnet/src/dotnetcore/GxClasses/GxClasses.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
 		<TargetFrameworks>netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
@@ -30,6 +30,7 @@
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataMemory.cs" Link="Data\GXDataMemory.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataPostgreSQL.cs" Link="Data\GXDataPostgreSQL.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataSqlite.cs" Link="Data\GXDataSqlite.cs" />
+    <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataHana.cs" Link="Data\GXDataHana.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Diagnostics\Log.cs" Link="Diagnostics\Log.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Diagnostics\GXDebugManager.cs" Link="Diagnostics\GXDebugManager.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXLDAP.cs" Link="Domain\GXLDAP.cs" />
